requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==4.0.3
attrs==23.1.0
auth0-python==4.4.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.0
charset-normalizer==3.2.0
cryptography==41.0.4
Django==3.2.8
django-bitfield==2.2.0
django-cors-headers==3.10.0
django-csp==3.7
django-filter==23.3
django-geo==0.8
django-js-asset==2.1.0
django-leaflet==0.29.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
django-restql==0.15.3
django-separatedvaluesfield==0.6.0
django-soft-delete==0.9.21
django-timezone-field==6.0.1
django-uuidfield==0.5.0
djangorestframework==3.12.4
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.21.7
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
lxml==4.9.3
multidict==6.0.4
packaging==23.1
phonenumbers==8.13.21
psycopg2-binary==2.9.9
pycparser==2.20
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyPEG2==2.15.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
South==1.0.2
sqlparse==0.4.2
typing-extensions==3.10.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.6
yarl==1.9.2
